+++ aten/src/ATen/templates/RegisterCUDA.cpp
@@ -4,7 +4,7 @@
 
 #include <ATen/Type.h>
 #include <ATen/Context.h>
-#include <ATen/detail/VariableHooksInterface.h>
+#include <ATen/core/VariableHooksInterface.h>
 
 ${cuda_type_headers}
